Consolidate some common functions to core submodule
